# docs/swagger.yaml
basePath: /
definitions:
  internal_api_auth.Credentials:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  internal_api_user.User:
    properties:
      id:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for a CRUD and Auth project using Gin and Swagger.
  title: Go BO-BOILERPLATE
  version: "1.0"
paths:
  /api/user:
    post:
      consumes:
      - application/json
      description: Create a new user with username and password
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_api_user.User'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            type: string
      summary: Create a new user
  /api/users:
    get:
      description: Get list of users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_api_user.User'
            type: array
      summary: Get list of users
  /login:
    post:
      consumes:
      - application/json
      description: Login a user with username and password
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/internal_api_auth.Credentials'
      produces:
      - application/json
      responses:
        "200":
          description: Token
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Login a user
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user with a username and password
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/internal_api_auth.Credentials'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            type: string
        "400":
          description: User already exists
          schema:
            type: string
      summary: Register a new user
swagger: "2.0"